+++ setup_containers.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
-echo 'Building SearchMyData image...'
-docker build -t searchmydata-v.1.7 .
+echo 'Fetching SearchMyData image...'
+docker pull andruxa17/searchmydata:v.1.7
 echo 'Done'
 echo 'Fetching MongoDB image...'
 docker pull mongo
@@ -13,4 +13,4 @@ docker run --net searchmydata-net --name database -d mongo
 echo 'Done'
 echo 'Run The App container...'
 sleep 1
-docker run --net searchmydata-net --name client -it searchmydata-v.1.7
+docker run --net searchmydata-net --name client -it andruxa17/searchmydata:v.1.7